Move internal `base-tab-list`
