+++ packages/radix-ui-themes/src/components/_internal/base-tab-list.props.ts
@@ -1,7 +1,7 @@
-import { colorPropDef } from '../props/color.prop.js';
-import { highContrastPropDef } from '../props/high-contrast.prop.js';
+import { colorPropDef } from '../../props/color.prop.js';
+import { highContrastPropDef } from '../../props/high-contrast.prop.js';
 
-import type { PropDef } from '../props/prop-def.js';
+import type { PropDef } from '../../props/prop-def.js';
 
 const sizes = ['1', '2'] as const;
 const wrapValues = ['nowrap', 'wrap', 'wrap-reverse'] as const;
