+++ packages/radix-ui-themes/src/components/tabs.css
@@ -1,4 +1,4 @@
-@import './base-tab-list.css';
+@import './_internal/base-tab-list.css';
 
 .rt-TabsContent {
   position: relative;
