+++ packages/radix-ui-themes/src/components/tabs.props.ts
@@ -8,5 +8,5 @@ const tabsContentPropDefs = {
   ...asChildPropDef,
 };
 
-export { baseTabListPropDefs as tabsListPropDefs } from './base-tab-list.props.js';
+export { baseTabListPropDefs as tabsListPropDefs } from './_internal/base-tab-list.props.js';
 export { tabsRootPropDefs, tabsContentPropDefs };
